+++ src/corejavaassignment/assignment2/Rectangle.java
@@ -1,4 +1,4 @@
-package CoreJavaAssignment.Assignment2;
+package corejavaassignment.assignment2;
 
 public class Rectangle {
