+++ src/corejavaassignment/assignment2/Square.java
@@ -1,11 +1,12 @@
-package CoreJavaAssignment.Assignment2;
+package corejavaassignment.assignment2;
 
-public class Cylinder {
+public class Square {
 
-    public double area(double radius, double height) {
+
+    public double area(double length) {
 
         try {
-            return Math.PI * Math.pow(radius, 2);
+            return Math.pow(length, 2);
         }
         catch (ArithmeticException arithmeticException){
             arithmeticException.printStackTrace();
@@ -13,14 +14,15 @@ public class Cylinder {
         }
     }
 
-    public double volume(double radius, double height) {
+    public double perimeter(double length) {
 
         try {
-            return Math.PI * Math.pow(radius, 2) * height;
+            return 4 * length;
         }
         catch (ArithmeticException arithmeticException){
             arithmeticException.printStackTrace();
             return 0.0;
         }
+
     }
 }
